requirements.txt
fastapi
uvicorn[standard]
chromadb
openai
python-dotenv
httpx
langchain-openai
sentence-transformers
pydantic
shapely
